+++ src/CollapsedCellOptimizer.cpp
@@ -298,7 +298,7 @@ void optimizeCell(SCExpT& experiment,
     char red[] = "\x1b[30m";
     red[3] = '0' + static_cast<char>(fmt::RED);
 
-    double cellCount {barcode};//numCells-jqueue.size_approx()};
+    double cellCount {static_cast<double>(barcode)};//numCells-jqueue.size_approx()};
     if (cellCount > totalCells) { cellCount = totalCells; }
     double percentCompletion {cellCount*100/numCells};
     fmt::print(stderr, "\033[A\r\r{}Analyzed {} cells ({}{}%{} of all).{}\n",
